requirements.txt
Faker==28.0.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
six==1.16.0
